# backend/openapi.json
{
  "components": {
    "schemas": {
      "AgentConfigResponse": {
        "properties": {
          "contract_version": {
            "const": 1,
            "default": 1,
            "title": "Contract Version",
            "type": "integer"
          },
          "max_session_seconds": {
            "title": "Max Session Seconds",
            "type": "integer"
          },
          "session": {
            "$ref": "#/components/schemas/AgentSession"
          },
          "warn_before_end_seconds": {
            "title": "Warn Before End Seconds",
            "type": "integer"
          }
        },
        "required": [
          "contract_version",
          "session",
          "max_session_seconds",
          "warn_before_end_seconds"
        ],
        "title": "AgentConfigResponse",
        "type": "object"
      },
...
      "AskPageRequest": {
        "additionalProperties": false,
        "properties": {
          "earlier_page_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 9,
            "title": "Earlier Page Ids",
            "type": "array"
          },
          "page_id": {
            "title": "Page Id",
            "type": "string"
          },
          "pointer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PointerHint"
              },
              {
                "type": "null"
              }
            ]
          },
          "question": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "page_id",
          "question"
        ],
        "title": "AskPageRequest",
        "type": "object"
      },
...
      "CreatePageRequest": {
        "additionalProperties": false,
        "properties": {
          "snapshot": {
            "$ref": "#/components/schemas/PageSnapshot"
          }
        },
        "required": [
          "snapshot"
        ],
        "title": "CreatePageRequest",
        "type": "object"
      },
      "CreatePageResponse": {
        "properties": {
          "contract_version": {
            "const": 1,
            "default": 1,
            "title": "Contract Version",
            "type": "integer"
          },
          "page_id": {
            "title": "Page Id",
            "type": "string"
          },
          "page_type": {
            "enum": [
              "terms",
              "privacy",
              "checkout",
              "form",
              "offer",
              "article",
              "other"
            ],
            "title": "Page Type",
            "type": "string"
          },
          "scan": {
            "$ref": "#/components/schemas/ScanBrief"
          },
          "site_signals": {
            "items": {
              "type": "string"
            },
            "title": "Site Signals",
            "type": "array"
          },
          "summary_for_agent": {
            "title": "Summary For Agent",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "page_id",
          "page_type",
          "summary_for_agent",
          "scan",
          "site_signals"
        ],
        "title": "CreatePageResponse",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "contract_version": {
            "const": 1,
            "default": 1,
            "title": "Contract Version",
            "type": "integer"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "contract_version",
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "ExplainFieldRequest": {
        "additionalProperties": false,
        "properties": {
          "field_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Id"
          },
          "field_label": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Label"
          },
          "page_id": {
            "title": "Page Id",
            "type": "string"
          },
          "pointer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PointerHint"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "page_id"
        ],
        "title": "ExplainFieldRequest",
        "type": "object"
      },
...
      "JourneySummaryRequest": {
        "additionalProperties": false,
        "description": "The trail's fact lines, oldest first, and the summary so far.",
        "properties": {
          "facts": {
            "items": {
              "maxLength": 600,
              "type": "string"
            },
            "maxItems": 10,
            "minItems": 1,
            "title": "Facts",
            "type": "array"
          },
          "previous_summary": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Summary"
          }
        },
        "required": [
          "facts"
        ],
        "title": "JourneySummaryRequest",
        "type": "object"
      },
      "JourneySummaryResponse": {
        "properties": {
          "contract_version": {
            "const": 1,
            "default": 1,
            "title": "Contract Version",
            "type": "integer"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "summary"
        ],
        "title": "JourneySummaryResponse",
        "type": "object"
      },
...
      "ScanRequest": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "anyOf": [
              {
                "enum": [
                  "terms",
                  "privacy",
                  "checkout",
                  "offer",
                  "general"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "page_id": {
            "title": "Page Id",
            "type": "string"
          }
        },
        "required": [
          "page_id"
        ],
        "title": "ScanRequest",
        "type": "object"
      },
      "ScanResult": {
        "properties": {
          "contract_version": {
            "const": 1,
            "default": 1,
            "title": "Contract Version",
            "type": "integer"
          },
          "counts": {
            "$ref": "#/components/schemas/ScanCounts"
          },
          "kind": {
            "anyOf": [
              {
                "enum": [
                  "terms",
                  "privacy",
                  "checkout",
                  "offer",
                  "general"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "nudge": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Nudge"
              },
              {
                "type": "null"
              }
            ]
          },
          "page_id": {
            "title": "Page Id",
            "type": "string"
          },
          "risks": {
            "items": {
              "$ref": "#/components/schemas/RiskFlag"
            },
            "title": "Risks",
            "type": "array"
          },
          "status": {
            "enum": [
              "pending",
              "ready",
              "failed",
              "none"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "page_id",
          "status",
          "kind",
          "risks",
          "counts",
          "nudge"
        ],
        "title": "ScanResult",
        "type": "object"
      },
...
      "SiteTrustRequest": {
        "additionalProperties": false,
        "properties": {
          "url": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "SiteTrustRequest",
        "type": "object"
      },
...
      "ToolResult": {
        "properties": {
          "agent_notes": {
            "title": "Agent Notes",
            "type": "string"
          },
          "card": {
            "anyOf": [
              {
                "discriminator": {
                  "mapping": {
                    "answer": "#/components/schemas/AnswerCard",
                    "field": "#/components/schemas/FieldCard",
                    "risk_list": "#/components/schemas/RiskListCard",
                    "summary": "#/components/schemas/SummaryCard",
                    "true_cost": "#/components/schemas/TrueCostCard",
                    "trust": "#/components/schemas/TrustCard",
                    "web_answer": "#/components/schemas/WebAnswerCard"
                  },
                  "propertyName": "kind"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/AnswerCard"
                  },
                  {
                    "$ref": "#/components/schemas/FieldCard"
                  },
                  {
                    "$ref": "#/components/schemas/RiskListCard"
                  },
                  {
                    "$ref": "#/components/schemas/SummaryCard"
                  },
                  {
                    "$ref": "#/components/schemas/WebAnswerCard"
                  },
                  {
                    "$ref": "#/components/schemas/TrueCostCard"
                  },
                  {
                    "$ref": "#/components/schemas/TrustCard"
                  }
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Card"
          },
          "contract_version": {
            "const": 1,
            "default": 1,
            "title": "Contract Version",
            "type": "integer"
          },
          "highlight_ids": {
            "items": {
              "type": "string"
            },
            "title": "Highlight Ids",
            "type": "array"
          },
          "not_found": {
            "title": "Not Found",
            "type": "boolean"
          },
          "quote_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quote Text"
          },
          "say": {
            "title": "Say",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "contract_version",
          "say",
          "agent_notes",
          "card",
          "highlight_ids",
          "quote_text",
          "sources",
          "not_found"
        ],
        "title": "ToolResult",
        "type": "object"
      },
...
      "WalkthroughResponse": {
        "properties": {
          "contract_version": {
            "const": 1,
            "default": 1,
            "title": "Contract Version",
            "type": "integer"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/WalkthroughStep"
            },
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "contract_version",
          "steps"
        ],
        "title": "WalkthroughResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Iris API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/agent-config": {
      "get": {
        "operationId": "agent_config_api_agent_config_get",
        "parameters": [
          {
            "in": "query",
            "name": "greet",
            "required": false,
            "schema": {
              "default": true,
              "title": "Greet",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "voice",
            "required": false,
            "schema": {
              "default": "female",
              "enum": [
                "female",
                "male"
              ],
              "title": "Voice",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Agent Config"
      }
    },
...
    "/api/journey/summary": {
      "post": {
        "operationId": "journey_summary_api_journey_summary_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JourneySummaryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneySummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Journey Summary"
      }
    },
    "/api/pages": {
      "post": {
        "operationId": "create_page_api_pages_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Create Page"
      }
    },
    "/api/pages/{page_id}/scan": {
      "get": {
        "operationId": "get_scan_api_pages__page_id__scan_get",
        "parameters": [
          {
            "in": "path",
            "name": "page_id",
            "required": true,
            "schema": {
              "title": "Page Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "exclude_keys",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 4000,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exclude Keys"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Get Scan"
      }
    },
    "/api/tools/ask-page": {
      "post": {
        "operationId": "ask_page_endpoint_api_tools_ask_page_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskPageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Ask Page Endpoint"
      }
    },
    "/api/tools/explain-field": {
      "post": {
        "operationId": "explain_field_endpoint_api_tools_explain_field_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplainFieldRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Explain Field Endpoint"
      }
    },
    "/api/tools/loan-cost": {
      "post": {
        "operationId": "loan_cost_endpoint_api_tools_loan_cost_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoanCostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Loan Cost Endpoint"
      }
    },
    "/api/tools/scan": {
      "post": {
        "operationId": "scan_endpoint_api_tools_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Scan Endpoint"
      }
    },
    "/api/tools/site-trust": {
      "post": {
        "operationId": "site_trust_endpoint_api_tools_site_trust_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteTrustRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Site Trust Endpoint"
      }
    },
    "/api/tools/summarize": {
      "post": {
        "operationId": "summarize_endpoint_api_tools_summarize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SummarizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Summarize Endpoint"
      }
    },
    "/api/tools/walkthrough": {
      "post": {
        "operationId": "walkthrough_endpoint_api_tools_walkthrough_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WalkthroughRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalkthroughResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Walkthrough Endpoint"
      }
    },
    "/api/tools/web-lookup": {
      "post": {
        "operationId": "web_lookup_endpoint_api_tools_web_lookup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebLookupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Web Lookup Endpoint"
      }
    },
    "/api/voice-token": {
      "get": {
        "operationId": "get_voice_token_api_voice_token_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Get Voice Token"
      }
    }
  }
}